Multigrid support added and could be enabled with --multigrid option
The multigrid option is not enabled by default because it is not enough stable for now but you can enable it with --multigrid option. Additionally the grids resize and clear functions are fixed and working as espected from now on. Mouse input needs some work on multigrid and floating windows are not correctly positioned.