Change project structure
main package is in cmd/neoray now because of go install command output binary name is package name, older one was generating an executable named "src.exe". Also packages are separated and moved to the pkg folder in root.
Font package removed and default fonts moved to the assets package in the  cmd/neoray/internal directory.